+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Added announcing status message for `zoom in` and `zoom out` button in `Preview` component.
+- Added `subTitle` to the `DataGridCell` component to display additional subtitle text in the row data.
 
 ### Fixed
 - Fixed accessibility issue with `DataGrid` component
@@ -12,6 +13,8 @@
 - Added `aria-erromessage` and `aria-describedby` to the `Autocomplete` component for improved screen reader accessibility.
 - Added a default `id` value to the `Autocomplete` component for making helper text accessible.
 - Fixed helper text rendering based on condition in `Select` component
+- Fixed screen reader accessibility issue with the pagination `label` in the `Pagination` component.
+- Fixed focus handling on the expand/collapse icon button in the `ProgressBar` component to address accessibility issues
 
 ### Changed